feat(checklist): add inventory reservation services
Rebuild reservations after family and inventory changes. Extract inventory matching, reservation rebuilds, checklist snapshot building, and inventory ID generation into focused services.
